% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{air}
\alias{air}
\title{The text of H.P. Lovecraft's "Cool air"}
\format{
A character vector with 339 elements
}
\source{
<https://www.gutenberg.org/ebooks/73177>
}
\usage{
air
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"Cool air".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
