% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{azathoth}
\alias{azathoth}
\title{"Azathoth"}
\format{
A character vector with 37 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
azathoth
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "Azathoth"
}
\examples{
data("azathoth")
lovecraftr::azathoth

}
\keyword{datasets}
