% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{beast_cave}
\alias{beast_cave}
\title{The text of H.P. Lovecraft's "The Beast in the Cave"}
\format{
A character vector with 153 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/bc.aspx>
}
\usage{
beast_cave
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Beast in the Cave".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
