% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beyond_the_wall_of_sleep}
\alias{beyond_the_wall_of_sleep}
\title{"Beyond the Wall of Sleep"}
\format{
A character vector with 361 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/beyond_the_wall_of_sleep}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/beyond_the_wall_of_sleep.rda}{RDA file}
}
}
\usage{
data(beyond_the_wall_of_sleep)
}
\description{
A dataset containing text of H. P. Lovecraft's "Beyond the Wall of Sleep".
}
\keyword{datasets}
\keyword{short}
