% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{book}
\alias{book}
\title{The text of H.P. Lovecraft's "The Book"}
\format{
A character vector with 75 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/b.aspx>
}
\usage{
book
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Book".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
