% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{call_of_cthulhu}
\alias{call_of_cthulhu}
\title{The text of H.P. Lovecraft's "The Call of Cthulhu"}
\format{
A character vector with 1194 elements
}
\source{
<https://www.gutenberg.org/ebooks/68283>
}
\usage{
call_of_cthulhu
}
\description{
A dataset containing the full text of H.P. Lovecraft's story
"The Call of Cthulhu".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
