% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{cats}
\alias{cats}
\title{The text of H.P. Lovecraft's "The Cats of Ulthar"}
\format{
A character vector with 82 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/cu.aspx>
}
\usage{
cats
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Cats of Ulthar".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
