% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celephais}
\alias{celephais}
\title{Dataset of "Celephais"}
\format{
A character vector with 202 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
celephais
}
\description{
A dataset containing text of H. P. Lovecraft's "Celephais"
}
\examples{
data("celephais")
lovecraftr::celephais

}
\keyword{datasets}
