% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{charles_dexter}
\alias{charles_dexter}
\title{The text of H.P. Lovecraft's "The Case of Charles Dexter Ward"}
\format{
A character vector with 4379 elements
}
\source{
<https://www.gutenberg.org/ebooks/73547>
}
\usage{
charles_dexter
}
\description{
A dataset containing the full text of H.P. Lovecraft's story
"The Case of Charles Dexter Ward".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
