% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{city}
\alias{city}
\title{The text of H.P. Lovecraft's "The Nameless City"}
\format{
A character vector with 299 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/nc.aspx>
}
\usage{
city
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Nameless City".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
