% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{colour_space}
\alias{colour_space}
\title{The text of H.P. Lovecraft's "The Colour Out of Space"}
\format{
A character vector with 1110 elements
}
\source{
<https://www.gutenberg.org/ebooks/68236>
}
\usage{
colour_space
}
\description{
A dataset containing the full text of H.P. Lovecraft's story
"The Colour Out of Space".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
