% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cool_air}
\alias{cool_air}
\title{Dataset of "COOL AIR"}
\format{
A character vector with 292 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/cool_air}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/cool_air.rda}{RDA file}.
}
\usage{
data(cool_air)
}
\description{
A dataset containing text of H. P. Lovecraft's "COOL AIR"
}
\keyword{datasets}
