% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dagon}
\alias{dagon}
\title{Dataset of "DAGON"}
\format{
A character vector with 187 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
dagon
}
\description{
A dataset containing text of H. P. Lovecraft's "DAGON"
}
\examples{
data("dagon")
lovecraftr::dagon

}
\keyword{datasets}
