% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{descendant}
\alias{descendant}
\title{The text of H.P. Lovecraft's "The Descendant"}
\format{
A character vector with 92 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/de.aspx>
}
\usage{
descendant
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Descendant".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
