% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{doom}
\alias{doom}
\title{The text of H.P. Lovecraft's "The Doom That Came to Sarnath"}
\format{
A character vector with 160 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/ds.aspx>
}
\usage{
doom
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Doom That Came to Sarnath".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
