% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{door_step}
\alias{door_step}
\title{The text of H.P. Lovecraft's "The thing on the door-step"}
\format{
A character vector with 1081 elements
}
\source{
<https://www.gutenberg.org/ebooks/73230>
}
\usage{
door_step
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The thing on the door-step".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
