% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dreams_in_the_witch_house}
\alias{dreams_in_the_witch_house}
\title{Dataset of "DREAMS IN THE WITCH-HOUSE"}
\format{
A character vector with 1,271 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/dreams_in_the_witch_house}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/dreams_in_the_witch_house.rda}{RDA file}.
}
\usage{
data(dreams_in_the_witch_house)
}
\description{
A dataset containing text of H. P. Lovecraft's "DREAMS IN THE WITCH-HOUSE"
}
\keyword{datasets}
