% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{dunwich_horror}
\alias{dunwich_horror}
\title{The text of H.P. Lovecraft's "The Dunwich Horror"}
\format{
A character vector with 1694 elements
}
\source{
<https://www.gutenberg.org/ebooks/50133>
}
\usage{
dunwich_horror
}
\description{
A dataset containing the full text of H.P. Lovecraft's story
"The Dunwich Horror".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
