% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{erich_zann}
\alias{erich_zann}
\title{The text of H.P. Lovecraft's "The Music of Erich Zann"}
\format{
A character vector with 198 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/mez.aspx>
}
\usage{
erich_zann
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Music of Erich Zann".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
