% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_oblivione}
\alias{ex_oblivione}
\title{Dataset of "EX OBLIVIONE"}
\format{
A character vector with 66 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
ex_oblivione
}
\description{
A dataset containing text of H. P. Lovecraft's "EX OBLIVIONE"
}
\examples{
data("ex_oblivione")
lovecraftr::ex_oblivione

}
\keyword{datasets}
