% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facts_concerning_the_late_arthur_jermyn_and_his_family}
\alias{facts_concerning_the_late_arthur_jermyn_and_his_family}
\title{Dataset of "FACTS CONCERNING THE LATE ARTHUR JERMYN AND HIS FAMILY"}
\format{
A character vector with 303 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
facts_concerning_the_late_arthur_jermyn_and_his_family
}
\description{
A dataset containing text of H. P. Lovecraft's "FACTS CONCERNING THE LATE ARTHUR JERMYN AND HIS FAMILY"
}
\examples{
data("facts_concerning_the_late_arthur_jermyn_and_his_family")
lovecraftr::facts_concerning_the_late_arthur_jermyn_and_his_family

}
\keyword{datasets}
