% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{festival}
\alias{festival}
\title{The text of H.P. Lovecraft's "The festival"}
\format{
A character vector with 345 elements
}
\source{
<https://www.gutenberg.org/ebooks/68553>
}
\usage{
festival
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The festival".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
