% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{haunter_dark}
\alias{haunter_dark}
\title{The text of H.P. Lovecraft's "The haunter of the dark"}
\format{
A character vector with 954 elements
}
\source{
<https://www.gutenberg.org/ebooks/73233>
}
\usage{
haunter_dark
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The haunter of the dark".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
