% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{he}
\alias{he}
\title{"He"}
\format{
A character vector with 360 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
he
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "He"
}
\examples{
data("he")
lovecraftr::he

}
\keyword{datasets}
