% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{hound}
\alias{hound}
\title{The text of H.P. Lovecraft's "The Hound"}
\format{
A character vector with 189 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/h.aspx>
}
\usage{
hound
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Hound".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
