% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypnos}
\alias{hypnos}
\title{"Hypnos"}
\format{
A character vector with 233 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
hypnos
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "Hypnos"
}
\examples{
data("hypnos")
lovecraftr::hypnos

}
\keyword{datasets}
