% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imprisoned_with_the_pharoahs}
\alias{imprisoned_with_the_pharoahs}
\title{"Imprisoned with the Pharaohs"}
\format{
A character vector with 948 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/imprisoned_with_the_pharoahs}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/imprisoned_with_the_pharoahs.rda}{RDA file}
}
}
\usage{
data(imprisoned_with_the_pharoahs)
}
\description{
A dataset containing text of H. P. Lovecraft's "Imprisoned with the Pharaohs".
}
\keyword{datasets}
\keyword{short}
