% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{in_the_vault}
\alias{in_the_vault}
\title{"In the Vault"}
\format{
A character vector with 272 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/in_the_vault}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/in_the_vault.rda}{RDA file}
}
}
\usage{
data(in_the_vault)
}
\description{
A dataset containing text of H. P. Lovecraft's "In the Vault".
}
\keyword{datasets}
\keyword{short}
