% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{iranon}
\alias{iranon}
\title{The text of H.P. Lovecraft's "The quest of Iranon"}
\format{
A character vector with 258 elements
}
\source{
<https://www.gutenberg.org/ebooks/73182>
}
\usage{
iranon
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The quest of Iranon".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
