% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{key}
\alias{key}
\title{The text of H.P. Lovecraft's "The silver key"}
\format{
A character vector with 472 elements
}
\source{
<https://www.gutenberg.org/ebooks/70478>
}
\usage{
key
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The silver key".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
