% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lovecraft}
\alias{lovecraft}
\title{Lovecraft}
\format{
A tibble with 43,347 rows and 2 columns:
\describe{
  \item{title}{Title of the work}
  \item{text}{~80 characters of text from the title}
}
}
\source{
Public domain
}
\usage{
lovecraft
}
\description{
A dataset containing UTF-8 plain text of
  all the H. P. Lovecraft works available in {lovecraftr}
}
\examples{
data("lovecraft")
lovecraftr::lovecraft

}
\keyword{datasets}
