% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft.R
\name{lovecraft_corpus}
\alias{lovecraft_corpus}
\title{Lovecraft's Corpus}
\usage{
lovecraft_corpus()
}
\value{
A data frame of Lovecraft's corpus
}
\description{
See `?lovecraftr::lovecraft_corpus` for documentation
}
\examples{
lovecraft_corpus()
}
