% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraftr-package.R
\docType{package}
\name{lovecraftr-package}
\alias{lovecraftr}
\alias{lovecraftr-package}
\title{lovecraftr: A Collection of Lovecraftian Tales and Texts}
\description{
A curated collection of Howard Phillips Lovecraft's complete stories, collected for the purpose of text analysis.
}
\author{
\strong{Maintainer}: Ruff Sergej \email{serijnh@gmail.com} (\href{https://orcid.org/0009-0000-8264-6347}{ORCID})

}
\keyword{internal}
