% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{lurking}
\alias{lurking}
\title{The text of H.P. Lovecraft's "The lurking fear"}
\format{
A character vector with 791 elements
}
\source{
<https://www.gutenberg.org/ebooks/70486>
}
\usage{
lurking
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The lurking fear".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
