% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medusas_coil}
\alias{medusas_coil}
\title{Dataset of "MEDUSA'S COIL"}
\format{
A character vector with 1,443 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
medusas_coil
}
\description{
A dataset containing text of H. P. Lovecraft's "MEDUSA'S COIL"
}
\examples{
data("medusas_coil")
lovecraftr::medusas_coil

}
\keyword{datasets}
