% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{mountain_madness}
\alias{mountain_madness}
\title{The text of H.P. Lovecraft's "At the Mountains of Madness"}
\format{
A character vector with 4618 elements
}
\source{
<https://www.gutenberg.org/ebooks/70652>
}
\usage{
mountain_madness
}
\description{
A dataset containing the full text of H.P. Lovecraft's story
"At the Mountains of Madness".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
