% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyarlathotep}
\alias{nyarlathotep}
\title{"Nyarlathotep"}
\format{
A character vector with 97 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/nyarlathotep}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/nyarlathotep.rda}{RDA file}
}
}
\usage{
data(nyarlathotep)
}
\description{
A dataset containing text of H. P. Lovecraft's "Nyarlathotep".
}
\keyword{datasets}
\keyword{short}
