% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{outsider}
\alias{outsider}
\title{The text of H.P. Lovecraft's "The Outsider"}
\format{
A character vector with 211 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/o.aspx>
}
\usage{
outsider
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Outsider".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
