% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pickmans_model}
\alias{pickmans_model}
\title{"Pickmans Model"}
\format{
A character vector with 455 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
pickmans_model
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "Pickmans Model"
}
\examples{
data("pickmans_model")
lovecraftr::pickmans_model

}
\keyword{datasets}
