% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polaris}
\alias{polaris}
\title{Dataset of "POLARIS"}
\format{
A character vector with 129 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
polaris
}
\description{
A dataset containing text of H. P. Lovecraft's "POLARIS"
}
\examples{
data("polaris")
lovecraftr::polaris

}
\keyword{datasets}
