% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{reanimator}
\alias{reanimator}
\title{The text of H.P. Lovecraft's "Herbert West—Reanimator"}
\format{
A character vector with 739 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/hwr.aspx>
}
\usage{
reanimator
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"Herbert West—Reanimator".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
