% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{red_hook}
\alias{red_hook}
\title{The text of H.P. Lovecraft's "The Horror at Red Hook"}
\format{
A character vector with 873 elements
}
\source{
<https://www.gutenberg.org/ebooks/72966>
}
\usage{
red_hook
}
\description{
A dataset containing the full text of H.P. Lovecraft's story
"The Horror at Red Hook".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
