% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{shadow_innsmouth}
\alias{shadow_innsmouth}
\title{The text of H.P. Lovecraft's "The Shadow Over Innsmouth"}
\format{
A character vector with 1708 elements
}
\source{
<https://www.gutenberg.org/ebooks/73181>
}
\usage{
shadow_innsmouth
}
\description{
A dataset containing the full text of H.P. Lovecraft's story
"The Shadow Over Innsmouth".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
