% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{shadow_time}
\alias{shadow_time}
\title{The text of H.P. Lovecraft's "The Shadow out of Time"}
\format{
A character vector with 1528 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/sot.aspx>
}
\usage{
shadow_time
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Shadow out of Time".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
