% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{shunned_house}
\alias{shunned_house}
\title{The text of H.P. Lovecraft's "The Shunned House"}
\format{
A character vector with 1048 elements
}
\source{
<https://www.gutenberg.org/ebooks/31469>
}
\usage{
shunned_house
}
\description{
A dataset containing the full text of H.P. Lovecraft's story
"The Shunned House".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
