% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{temple}
\alias{temple}
\title{The text of H.P. Lovecraft's "The Temple"}
\format{
A character vector with 319 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/te.aspx>
}
\usage{
temple
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Temple".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
