% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_alchemist}
\alias{the_alchemist}
\title{"The Alchemist"}
\format{
A character vector with 295 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_alchemist
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Alchemist"
}
\examples{
data("the_alchemist")
lovecraftr::the_alchemist

}
\keyword{datasets}
