% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_beast_in_the_cave}
\alias{the_beast_in_the_cave}
\title{Dataset of "THE BEAST IN THE CAVE"}
\format{
A character vector with 204 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_beast_in_the_cave
}
\description{
A dataset containing text of H. P. Lovecraft's "THE BEAST IN THE CAVE"
}
\examples{
data("the_beast_in_the_cave")
lovecraftr::the_beast_in_the_cave

}
\keyword{datasets}
