% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_crawling_chaos}
\alias{the_crawling_chaos}
\title{Dataset of "THE CRAWLING CHAOS"}
\format{
A character vector with 234 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_crawling_chaos
}
\description{
A dataset containing text of H. P. Lovecraft's "THE CRAWLING CHAOS"
}
\examples{
data("the_crawling_chaos")
lovecraftr::the_crawling_chaos

}
\keyword{datasets}
