% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_doom_that_came_to_sarnath}
\alias{the_doom_that_came_to_sarnath}
\title{Dataset of "THE DOOM THAT CAME TO SARNATH"}
\format{
A character vector with 218 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_doom_that_came_to_sarnath
}
\description{
A dataset containing text of H. P. Lovecraft's "THE DOOM THAT CAME TO SARNATH"
}
\examples{
data("the_doom_that_came_to_sarnath")
lovecraftr::the_doom_that_came_to_sarnath

}
\keyword{datasets}
