% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_dream_quest_of_unknown_kadath}
\alias{the_dream_quest_of_unknown_kadath}
\title{Dataset of "THE DREAM QUEST OF UNKNOWN KADATH"}
\format{
A character vector with 3,475 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_dream_quest_of_unknown_kadath}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_dream_quest_of_unknown_kadath.rda}{RDA file}.
}
\usage{
data(the_dream_quest_of_unknown_kadath)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE DREAM QUEST OF UNKNOWN KADATH"
}
\keyword{datasets}
