% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_evil_clergyman}
\alias{the_evil_clergyman}
\title{"The Evil Clergyman"}
\format{
A character vector with 147 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_evil_clergyman}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_evil_clergyman.rda}{RDA file}
}
}
\usage{
data(the_evil_clergyman)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Evil Clergyman".
}
\keyword{datasets}
\keyword{short}
