% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_haunter_of_the_dark}
\alias{the_haunter_of_the_dark}
\title{"The Haunter Of The Dark"}
\format{
A character vector with 828 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_haunter_of_the_dark
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Haunter Of The Dark"
}
\examples{
data("the_haunter_of_the_dark")
lovecraftr::the_haunter_of_the_dark

}
\keyword{datasets}
