% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_horror_at_martins_beach}
\alias{the_horror_at_martins_beach}
\title{"The Horror At Martins Beach"}
\format{
A character vector with 225 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_horror_at_martins_beach
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Horror At Martins Beach"
}
\examples{
data("the_horror_at_martins_beach")
lovecraftr::the_horror_at_martins_beach

}
\keyword{datasets}
