% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_hound}
\alias{the_hound}
\title{Dataset of "THE HOUND"}
\format{
A character vector with 266 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_hound
}
\description{
A dataset containing text of H. P. Lovecraft's "THE HOUND"
}
\examples{
data("the_hound")
lovecraftr::the_hound

}
\keyword{datasets}
