% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_moon_bog}
\alias{the_moon_bog}
\title{Dataset of "The Moon Bog"}
\format{
A character vector with 266 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_moon_bog
}
\description{
A dataset containing text of H. P. Lovecraft's "The Moon Bog"
}
\examples{
data("the_moon_bog")
lovecraftr::the_moon_bog

}
\keyword{datasets}
