% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_other_gods}
\alias{the_other_gods}
\title{Dataset of "THE OTHER GODS"}
\format{
A character vector with 160 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_other_gods}{raw text file} or download
the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_other_gods.rda}{RDA file} to use with R.
}
\usage{
data(the_other_gods)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE OTHER GODS"
}
\examples{
\dontrun{
library(lovecraftr)
data(the_other_gods)

the_other_gods <- lovecraftr::the_other_gods
}
}
\keyword{datasets}
