% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_outsider}
\alias{the_outsider}
\title{"The Outsider"}
\format{
A character vector with 213 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_outsider}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_outsider.rda}{RDA file}
}
}
\usage{
data(the_outsider)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Outsider".
}
\keyword{datasets}
\keyword{short}
