% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_quest_of_iranon}
\alias{the_quest_of_iranon}
\title{"The Quest of Iranon"}
\format{
A character vector with 215 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_quest_of_iranon}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_quest_of_iranon.rda}{RDA file}
}
}
\usage{
data(the_quest_of_iranon)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Quest of Iranon".
}
\keyword{datasets}
\keyword{short}
