% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_rats_in_the_walls}
\alias{the_rats_in_the_walls}
\title{"The Rats In The Walls"}
\format{
A character vector with 691 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_rats_in_the_walls
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Rats In The Walls"
}
\examples{
data("the_rats_in_the_walls")
lovecraftr::the_rats_in_the_walls

}
\keyword{datasets}
