% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_shadow_out_of_time}
\alias{the_shadow_out_of_time}
\title{Dataset of \emph{The Shadow Out of Time}}
\format{
A character vector with 2,529 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_shadow_out_of_time}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_shadow_out_of_time.rda}{RDA file}.
}
\usage{
data(the_shadow_out_of_time)
}
\description{
A dataset containing text of H. P. Lovecraft's \emph{The Shadow Out of Time}
}
\keyword{datasets}
\keyword{novella}
