% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_shadow_over_innsmouth}
\alias{the_shadow_over_innsmouth}
\title{Dataset of "THE SHADOW OVER INNSMOUTH"}
\format{
A character vector with 2,318 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_shadow_over_innsmouth
}
\description{
A dataset containing text of H. P. Lovecraft's "THE SHADOW OVER INNSMOUTH"
}
\examples{
data("the_shadow_over_innsmouth")
lovecraftr::the_shadow_over_innsmouth

}
\keyword{datasets}
