% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_shunned_house}
\alias{the_shunned_house}
\title{"The Shunned House"}
\format{
A character vector with 916 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_shunned_house
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Shunned House"
}
\examples{
data("the_shunned_house")
lovecraftr::the_shunned_house

}
\keyword{datasets}
