% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_silver_key}
\alias{the_silver_key}
\title{Dataset of "THE SILVER KEY"}
\format{
A character vector with 404 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_silver_key}{raw text file} or download
the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_silver_key.rda}{RDA file} to use with R.
}
\usage{
data(the_silver_key)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE SILVER KEY"
}
\examples{
\dontrun{
library(lovecraftr)
data(the_silver_key)

the_silver_key <- lovecraftr::the_silver_key
}
}
\keyword{datasets}
