% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_statement_of_randolph_carter}
\alias{the_statement_of_randolph_carter}
\title{Dataset of "THE STATEMENT OF RANDOLPH CARTER"}
\format{
A character vector with 221 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_statement_of_randolph_carter
}
\description{
A dataset containing text of H. P. Lovecraft's "THE STATEMENT OF RANDOLPH CARTER"
}
\examples{
data("the_statement_of_randolph_carter")
lovecraftr::the_statement_of_randolph_carter

}
\keyword{datasets}
