% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_strange_high_house_in_the_mist}
\alias{the_strange_high_house_in_the_mist}
\title{Dataset of "THE STRANGE HIGH HOUSE IN THE MIST"}
\format{
A character vector with 289 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_strange_high_house_in_the_mist
}
\description{
A dataset containing text of H. P. Lovecraft's "THE STRANGE HIGH HOUSE IN THE MIST"
}
\examples{
data("the_strange_high_house_in_the_mist")
lovecraftr::the_strange_high_house_in_the_mist

}
\keyword{datasets}
