% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_temple}
\alias{the_temple}
\title{Dataset of "The Temple"}
\format{
A character vector with 450 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_temple
}
\description{
A dataset containing text of H. P. Lovecraft's "The Temple"
}
\examples{
data("the_temple")
lovecraftr::the_temple

}
\keyword{datasets}
