% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_terrible_old_man}
\alias{the_terrible_old_man}
\title{Dataset of "The Terrible Old Man"}
\format{
A character vector with 87 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_terrible_old_man
}
\description{
A dataset containing text of H. P. Lovecraft's "The Terrible Old Man"
}
\examples{
data("the_terrible_old_man")
lovecraftr::the_terrible_old_man

}
\keyword{datasets}
