% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_thing_on_the_doorstep}
\alias{the_thing_on_the_doorstep}
\title{"The Thing On The Doorstep"}
\format{
A character vector with 931 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_thing_on_the_doorstep
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Thing On The Doorstep"
}
\examples{
data("the_thing_on_the_doorstep")
lovecraftr::the_thing_on_the_doorstep

}
\keyword{datasets}
