% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_tomb}
\alias{the_tomb}
\title{"The Tomb"}
\format{
A character vector with 366 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_tomb}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_tomb.rda}{RDA file}
}
}
\usage{
data(the_tomb)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Tomb".
}
\keyword{datasets}
\keyword{short}
