% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_transition_of_juan_romero}
\alias{the_transition_of_juan_romero}
\title{"The Transition of Juan Romero"}
\format{
A character vector with 234 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_transition_of_juan_romero}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_transition_of_juan_romero.rda}{RDA file}
}
}
\usage{
data(the_transition_of_juan_romero)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Transition of Juan Romero".
}
\keyword{datasets}
\keyword{short}
