% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_tree}
\alias{the_tree}
\title{"The Tree"}
\format{
A character vector with 127 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_tree
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Tree"
}
\examples{
data("the_tree")
lovecraftr::the_tree

}
\keyword{datasets}
