% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_unnamable}
\alias{the_unnamable}
\title{Dataset of "The Unnamable"}
\format{
A character vector with 239 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_unnamable
}
\description{
A dataset containing text of H. P. Lovecraft's "The Unnamable"
}
\examples{
data("the_unnamable")
lovecraftr::the_unnamable

}
\keyword{datasets}
