% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{through_the_gates_of_the_silver_key}
\alias{through_the_gates_of_the_silver_key}
\title{Dataset of "THROUGH THE GATES OF THE SILVER KEY"}
\format{
A character vector with 1,305 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
through_the_gates_of_the_silver_key
}
\description{
A dataset containing text of H. P. Lovecraft's "THROUGH THE GATES OF THE SILVER KEY"
}
\examples{
data("through_the_gates_of_the_silver_key")
lovecraftr::through_the_gates_of_the_silver_key

}
\keyword{datasets}
