% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{united_amateur}
\alias{united_amateur}
\title{The text of H.P. Lovecraft's "Writings in the United Amateur"}
\format{
A character vector with 10749 elements
}
\source{
<https://www.gutenberg.org/ebooks/30637>
}
\usage{
united_amateur
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"Writings in the United Amateur".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
