% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{unknown_kadath}
\alias{unknown_kadath}
\title{The text of H.P. Lovecraft's "The Dream-Quest of Unknown Kadath"}
\format{
A character vector with 2483 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/dq.aspx>
}
\usage{
unknown_kadath
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Dream-Quest of Unknown Kadath".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
