% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{wall_sleep}
\alias{wall_sleep}
\title{The text of H.P. Lovecraft's "Beyond the Wall of Sleep"}
\format{
A character vector with 274 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/bws.aspx>
}
\usage{
wall_sleep
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"Beyond the Wall of Sleep".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
