% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{witch_house}
\alias{witch_house}
\title{The text of H.P. Lovecraft's "The Dreams in the Witch House"}
\format{
A character vector with 892 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/dwh.aspx>
}
\usage{
witch_house
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"The Dreams in the Witch House".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
