% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{at_the_mountains_of_madness}
\alias{at_the_mountains_of_madness}
\title{\emph{At the Mountains of Madness}}
\format{
A character vector with 3,579 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/at_the_mountains_of_madness}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/at_the_mountains_of_madness.rda}{RDA file}.
}
\usage{
data(at_the_mountains_of_madness)
}
\description{
A dataset containing text of H. P. Lovecraft's \emph{At the Mountains of Madness}.
}
\keyword{datasets}
\keyword{novella}
