% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{at_the_mountains_of_madness}
\alias{at_the_mountains_of_madness}
\title{"At The Mountains Of Madness"}
\format{
A character vector with 3,579 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
at_the_mountains_of_madness
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "At The Mountains Of Madness"
}
\examples{
data("at_the_mountains_of_madness")
lovecraftr::at_the_mountains_of_madness

}
\keyword{datasets}
