% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{azathoth}
\alias{azathoth}
\title{The text of H.P. Lovecraft's "Azathoth"}
\format{
A character vector with 34 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/az.aspx>
}
\usage{
azathoth
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"Azathoth".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
