% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{azathoth}
\alias{azathoth}
\title{"Azathoth"}
\format{
A character vector with 37 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/azathoth}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/azathoth.rda}{RDA file}
}
}
\usage{
data(azathoth)
}
\description{
A dataset containing text of H. P. Lovecraft's "Azathoth".
}
\keyword{datasets}
\keyword{short}
