% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{azathoth}
\alias{azathoth}
\title{Dataset of "AZATHOTH"}
\format{
A character vector with 37 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
azathoth
}
\description{
A dataset containing text of H. P. Lovecraft's "AZATHOTH"
}
\examples{
data("azathoth")
lovecraftr::azathoth

}
\keyword{datasets}
