% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{azathoth}
\alias{azathoth}
\title{Dataset of "AZATHOTH"}
\format{
A character vector with 37 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/azathoth}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/azathoth.rda}{RDA file}.
}
\usage{
data(azathoth)
}
\description{
A dataset containing text of H. P. Lovecraft's "AZATHOTH"
}
\keyword{datasets}
