% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beyond_the_wall_of_sleep}
\alias{beyond_the_wall_of_sleep}
\title{"Beyond The Wall Of Sleep"}
\format{
A character vector with 361 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
beyond_the_wall_of_sleep
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "Beyond The Wall Of Sleep"
}
\examples{
data("beyond_the_wall_of_sleep")
lovecraftr::beyond_the_wall_of_sleep

}
\keyword{datasets}
