% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beyond_the_wall_of_sleep}
\alias{beyond_the_wall_of_sleep}
\title{Dataset of "Beyond the Wall of Sleep"}
\format{
A character vector with 361 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/beyond_the_wall_of_sleep}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/beyond_the_wall_of_sleep.rda}{RDA file}.
}
\usage{
data(beyond_the_wall_of_sleep)
}
\description{
A dataset containing text of H. P. Lovecraft's "Beyond the Wall of Sleep"
}
\keyword{datasets}
\keyword{short}
