% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{celephais}
\alias{celephais}
\title{The text of H.P. Lovecraft's "Celephaïs"}
\format{
A character vector with 153 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/c.aspx>
}
\usage{
celephais
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"Celephaïs".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
