% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cool_air}
\alias{cool_air}
\title{Dataset of "Cool Air"}
\format{
A character vector with 292 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
cool_air
}
\description{
A dataset containing text of H. P. Lovecraft's "Cool Air"
}
\examples{
data("cool_air")
lovecraftr::cool_air

}
\keyword{datasets}
