% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{dagon}
\alias{dagon}
\title{The text of H.P. Lovecraft's "Dagon"}
\format{
A character vector with 133 elements
}
\source{
<https://www.hplovecraft.com/writings/texts/fiction/d.aspx>
}
\usage{
dagon
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"Dagon".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
