% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dagon}
\alias{dagon}
\title{Dataset of "Dagon"}
\format{
A character vector with 187 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/dagon}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/dagon.rda}{RDA file}.
}
\usage{
data(dagon)
}
\description{
A dataset containing text of H. P. Lovecraft's "Dagon"
}
\keyword{datasets}
\keyword{short}
