% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dreams_in_the_witch_house}
\alias{dreams_in_the_witch_house}
\title{Dataset of "Dreams In The Witch-House"}
\format{
A character vector with 1,271 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
dreams_in_the_witch_house
}
\description{
A dataset containing text of H. P. Lovecraft's "Dreams In The Witch-House"
}
\examples{
data("dreams_in_the_witch_house")
lovecraftr::dreams_in_the_witch_house

}
\keyword{datasets}
