% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dreams_in_the_witch_house}
\alias{dreams_in_the_witch_house}
\title{"Dreams in the Witch-House"}
\format{
A character vector with 1,271 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/dreams_in_the_witch_house}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/dreams_in_the_witch_house.rda}{RDA file}
}
}
\usage{
data(dreams_in_the_witch_house)
}
\description{
A dataset containing text of H. P. Lovecraft's "Dreams in the Witch-House".
}
\keyword{datasets}
\keyword{short}
