% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_oblivione}
\alias{ex_oblivione}
\title{Dataset of "Ex Oblivione"}
\format{
A character vector with 66 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/ex_oblivione}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/ex_oblivione.rda}{RDA file}.
}
\usage{
data(ex_oblivione)
}
\description{
A dataset containing text of H. P. Lovecraft's "Ex Oblivione"
}
\keyword{datasets}
\keyword{short}
