% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facts_concerning_the_late_arthur_jermyn_and_his_family}
\alias{facts_concerning_the_late_arthur_jermyn_and_his_family}
\title{"Facts Concerning the Late Arthur Jermyn and His Family"}
\format{
A character vector with 303 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/facts_concerning_the_late_arthur_jermyn_and_his_family}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/facts_concerning_the_late_arthur_jermyn_and_his_family.rda}{RDA file}
}
}
\usage{
data(facts_concerning_the_late_arthur_jermyn_and_his_family)
}
\description{
A dataset containing text of H. P. Lovecraft's "Facts Concerning the Late Arthur Jermyn and His Family".
}
\keyword{datasets}
\keyword{short}
