% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{from_beyond}
\alias{from_beyond}
\title{Dataset of "FROM BEYOND"}
\format{
A character vector with 222 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
from_beyond
}
\description{
A dataset containing text of H. P. Lovecraft's "FROM BEYOND"
}
\examples{
data("from_beyond")
lovecraftr::from_beyond

}
\keyword{datasets}
