% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{from_beyond}
\alias{from_beyond}
\title{Dataset of "From Beyond"}
\format{
A character vector with 222 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/from_beyond}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/from_beyond.rda}{RDA file}.
}
\usage{
data(from_beyond)
}
\description{
A dataset containing text of H. P. Lovecraft's "From Beyond"
}
\keyword{datasets}
\keyword{short}
