% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft_stories.R
\docType{data}
\name{he}
\alias{he}
\title{The text of H.P. Lovecraft's "He"}
\format{
A character vector with 419 elements
}
\source{
<https://www.gutenberg.org/ebooks/68547>
}
\usage{
he
}
\description{
A dataset containing the full text of H.P. Lovecraft's
"He".
The text is provided as a character vector, with each element
representing a chunk of the full text.
}
\keyword{datasets}
