% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{he}
\alias{he}
\title{Dataset of "He"}
\format{
A character vector with 360 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/he}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/he.rda}{RDA file}.
}
\usage{
data(he)
}
\description{
A dataset containing text of H. P. Lovecraft's "He"
}
\keyword{datasets}
\keyword{short}
