% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{herbert_west_reanimator}
\alias{herbert_west_reanimator}
\title{Dataset of "Herbert West: Reanimator"}
\format{
A character vector with 1,053 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/herbert_west_reanimator}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/herbert_west_reanimator.rda}{RDA file}.
}
\usage{
data(herbert_west_reanimator)
}
\description{
A dataset containing text of H. P. Lovecraft's "Herbert West: Reanimator"
}
\keyword{datasets}
\keyword{short}
