% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{herbert_west_reanimator}
\alias{herbert_west_reanimator}
\title{"Herbert West: Reanimator"}
\format{
A character vector with 1,053 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
herbert_west_reanimator
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "Herbert West: Reanimator"
}
\examples{
data("herbert_west_reanimator")
lovecraftr::herbert_west_reanimator

}
\keyword{datasets}
