% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypnos}
\alias{hypnos}
\title{Dataset of "HYPNOS"}
\format{
A character vector with 233 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/hypnos}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/hypnos.rda}{RDA file}.
}
\usage{
data(hypnos)
}
\description{
A dataset containing text of H. P. Lovecraft's "HYPNOS"
}
\keyword{datasets}
