% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imprisoned_with_the_pharoahs}
\alias{imprisoned_with_the_pharoahs}
\title{Dataset of "IMPRISONED WITH THE PHARAOHS"}
\format{
A character vector with 948 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
imprisoned_with_the_pharoahs
}
\description{
A dataset containing text of H. P. Lovecraft's "IMPRISONED WITH THE PHARAOHS"
}
\examples{
data("imprisoned_with_the_pharoahs")
lovecraftr::imprisoned_with_the_pharoahs

}
\keyword{datasets}
