% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{in_the_vault}
\alias{in_the_vault}
\title{Dataset of "IN THE VAULT"}
\format{
A character vector with 272 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
in_the_vault
}
\description{
A dataset containing text of H. P. Lovecraft's "IN THE VAULT"
}
\examples{
data("in_the_vault")
lovecraftr::in_the_vault

}
\keyword{datasets}
