% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraft.R
\name{lovecraft_corpus}
\alias{lovecraft_corpus}
\title{Lovecraft's Corpus}
\usage{
lovecraft_corpus()
}
\value{
A tibble of Lovecraft's corpus.
}
\description{
See \code{\link[lovecraftr:lovecraft]{lovecraftr:lovecraft}} for documentation.
}
\seealso{
\code{\link[lovecraftr:lovecraft]{lovecraftr:lovecraft}}
}
