% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lovecraftr-package.R
\docType{package}
\name{lovecraftr}
\alias{lovecraftr}
\alias{lovecraftr-package}
\title{Lovecraftr: A Collection of H.P. Lovecraft's Classic Stories}
\description{
The \emph{lovecraftr} package offers a collection of H.P. Lovecraft's most renowned stories,
formatted for easy access and text analysis. This package is intended for literary study,
text analysis, and exploration of Lovecraft’s unique themes and style within the horror genre.

List of Stories Included in the \emph{lovecraftr} Package:

\describe{
  \item{\emph{At the Mountains of Madness} - MM}{}
  \item{\emph{Azathoth} - AZA}{}
  \item{\emph{Beyond the Wall of Sleep} - BWS}{}
  \item{\emph{Celephais} - C}{}
  \item{\emph{Cool Air} - air}{}
  \item{\emph{Dagon} - D}{}
  \item{\emph{He} - HE}{}
  \item{\emph{Herbert West—Reanimator} - HWR}{}
  \item{\emph{The Alchemist} - ALC}{}
  \item{\emph{The Beast in the Cave} - BIC}{}
  \item{\emph{The Book} - book}{}
  \item{\emph{The Call of Cthulhu} - CC}{}
  \item{\emph{The Case of Charles Dexter Ward} - CDW}{}
  \item{\emph{The Cats of Ulthar} - cat}{}
  \item{\emph{The Colour Out of Space} - CS}{}
  \item{\emph{The Descendant} - DE}{}
  \item{\emph{The Doom That Came to Sarnath} - DS}{}
  \item{\emph{The Dreams in the Witch House} - DWH}{}
  \item{\emph{The Dream-Quest of Unknown Kadath} - DQ}{}
  \item{\emph{The Dunwich Horror} - DW}{}
  \item{\emph{The Festival} - FE}{}
  \item{\emph{The Haunter of the Dark} - haunter}{}
  \item{\emph{The Horror at Red Hook} - RH}{}
  \item{\emph{The Hound} - H}{}
  \item{\emph{The Lurking Fear} - LURK}{}
  \item{\emph{The Music of Erich Zann} - MEZ}{}
  \item{\emph{The Nameless City} - NC}{}
  \item{\emph{The Outsider} - O}{}
  \item{\emph{The Quest of Iranon} - IRA}{}
  \item{\emph{The Shadow out of Time} - SOT}{}
  \item{\emph{The Shadow over Innsmouth} - SOI}{}
  \item{\emph{The Shunned House} - TSH}{}
  \item{\emph{The Silver Key} - key}{}
  \item{\emph{The Temple} - TE}{}
  \item{\emph{The Thing on the Doorstep} - door}{}
  \item{\emph{Writings in the United Amateur} - WUA}{}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SergejRuff/lovecraftr}
  \item Report bugs at \url{https://github.com/SergejRuff/lovecraftr/issues}
}

}
\author{
\strong{Maintainer}: Sergej Ruff (\email{serijnh@gmail.com})

If you have any questions, suggestions, or issues, please feel free to contact me (\email{serijnh@gmail.com}).
}
