% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memory}
\alias{memory}
\title{Dataset of "Memory"}
\format{
A character vector with 31 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
memory
}
\description{
A dataset containing text of H. P. Lovecraft's "Memory"
}
\examples{
data("memory")
lovecraftr::memory

}
\keyword{datasets}
