% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memory}
\alias{memory}
\title{Dataset of "Memory"}
\format{
A character vector with 31 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/memory}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/memory.rda}{RDA file}.
}
\usage{
data(memory)
}
\description{
A dataset containing text of H. P. Lovecraft's "Memory"
}
\keyword{datasets}
\keyword{short}
