% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyarlathotep}
\alias{nyarlathotep}
\title{Dataset of "Nyarlathotep"}
\format{
A character vector with 97 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/nyarlathotep}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/nyarlathotep.rda}{RDA file}.
}
\usage{
data(nyarlathotep)
}
\description{
A dataset containing text of H. P. Lovecraft's "Nyarlathotep"
}
\keyword{datasets}
\keyword{short}
