% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyarlathotep}
\alias{nyarlathotep}
\title{Dataset of "Nyarlathotep"}
\format{
A character vector with 97 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
nyarlathotep
}
\description{
A dataset containing text of H. P. Lovecraft's "Nyarlathotep"
}
\examples{
data("nyarlathotep")
lovecraftr::nyarlathotep

}
\keyword{datasets}
