% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pickmans_model}
\alias{pickmans_model}
\title{"Pickman's Model"}
\format{
A character vector with 455 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/pickmans_model}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/pickmans_model.rda}{RDA file}.
}
\usage{
data(pickmans_model)
}
\description{
A dataset containing text of H. P. Lovecraft's "Pickman's Model".
}
\keyword{datasets}
\keyword{short}
