% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pickmans_model}
\alias{pickmans_model}
\title{Dataset of "Pickman's Model"}
\format{
A character vector with 455 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
pickmans_model
}
\description{
A dataset containing text of H. P. Lovecraft's "Pickman's Model"
}
\examples{
data("pickmans_model")
lovecraftr::pickmans_model

}
\keyword{datasets}
