% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poetry_of_the_gods}
\alias{poetry_of_the_gods}
\title{Dataset of "Poetry Of The Gods"}
\format{
A character vector with 157 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
poetry_of_the_gods
}
\description{
A dataset containing text of H. P. Lovecraft's "Poetry Of The Gods"
}
\examples{
data("poetry_of_the_gods")
lovecraftr::poetry_of_the_gods

}
\keyword{datasets}
