% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poetry_of_the_gods}
\alias{poetry_of_the_gods}
\title{Dataset of "Poetry of the Gods"}
\format{
A character vector with 157 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/poetry_of_the_gods}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/poetry_of_the_gods.rda}{RDA file}.
}
\usage{
data(poetry_of_the_gods)
}
\description{
A dataset containing text of H. P. Lovecraft's "Poetry of the Gods"
}
\keyword{datasets}
\keyword{short}
