% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polaris}
\alias{polaris}
\title{"Polaris"}
\format{
A character vector with 129 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
polaris
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "Polaris"
}
\examples{
data("polaris")
lovecraftr::polaris

}
\keyword{datasets}
