% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polaris}
\alias{polaris}
\title{Dataset of "Polaris"}
\format{
A character vector with 129 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/polaris}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/polaris.rda}{RDA file}.
}
\usage{
data(polaris)
}
\description{
A dataset containing text of H. P. Lovecraft's "Polaris"
}
\keyword{datasets}
\keyword{short}
