% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_alchemist}
\alias{the_alchemist}
\title{Dataset of "THE ALCHEMIST"}
\format{
A character vector with 295 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_alchemist}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_alchemist.rda}{RDA file}.
}
\usage{
data(the_alchemist)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE ALCHEMIST"
}
\keyword{datasets}
