% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_book}
\alias{the_book}
\title{"The Book"}
\format{
A character vector with 95 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_book
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Book"
}
\examples{
data("the_book")
lovecraftr::the_book

}
\keyword{datasets}
