% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_call_of_cthulhu}
\alias{the_call_of_cthulhu}
\title{Dataset of "THE CALL OF CTHULHU"}
\format{
A character vector with 1,030 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_call_of_cthulhu}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_call_of_cthulhu.rda}{RDA file}.
}
\usage{
data(the_call_of_cthulhu)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE CALL OF CTHULHU"
}
\keyword{datasets}
