% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_case_of_charles_dexter_ward}
\alias{the_case_of_charles_dexter_ward}
\title{Dataset of \emph{The Case of Charles Dexter Ward}}
\format{
A character vector with 4,358 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_case_of_charles_dexter_ward}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_case_of_charles_dexter_ward.rda}{RDA file}.
}
\usage{
data(the_case_of_charles_dexter_ward)
}
\description{
A dataset containing text of H. P. Lovecraft's \emph{The Case of Charles Dexter Ward}
}
\keyword{datasets}
\keyword{novella}
