% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_colour_out_of_space}
\alias{the_colour_out_of_space}
\title{"The Colour Out of Space"}
\format{
A character vector with 967 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_colour_out_of_space}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_colour_out_of_space.rda}{RDA file}
}
}
\usage{
data(the_colour_out_of_space)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Colour Out of Space".
}
\keyword{datasets}
\keyword{short}
