% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_colour_out_of_space}
\alias{the_colour_out_of_space}
\title{"The Colour Out Of Space"}
\format{
A character vector with 967 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_colour_out_of_space
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Colour Out Of Space"
}
\examples{
data("the_colour_out_of_space")
lovecraftr::the_colour_out_of_space

}
\keyword{datasets}
