% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_descendent}
\alias{the_descendent}
\title{Dataset of "THE DESCENDANT"}
\format{
A character vector with 121 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_descendent}{raw text file} or download
the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_descendent.rda}{RDA file} to use with R.
}
\usage{
data(the_descendent)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE DESCENDANT"
}
\examples{
\dontrun{
library(lovecraftr)
data(the_descendent)

the_descendent <- lovecraftr::the_descendent
}
}
\keyword{datasets}
