% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_descendent}
\alias{the_descendent}
\title{Dataset of "THE DESCENDANT"}
\format{
A character vector with 121 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_descendent
}
\description{
A dataset containing text of H. P. Lovecraft's "THE DESCENDANT"
}
\examples{
data("the_descendent")
lovecraftr::the_descendent

}
\keyword{datasets}
