% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_descendent}
\alias{the_descendent}
\title{"The Descendent"}
\format{
A character vector with 121 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_descendent
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Descendent"
}
\examples{
data("the_descendent")
lovecraftr::the_descendent

}
\keyword{datasets}
