% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_doom_that_came_to_sarnath}
\alias{the_doom_that_came_to_sarnath}
\title{"The Doom that Came to Sarnath"}
\format{
A character vector with 218 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_doom_that_came_to_sarnath}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_doom_that_came_to_sarnath.rda}{RDA file}.
}
\usage{
data(the_doom_that_came_to_sarnath)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Doom that Came to Sarnath".
}
\keyword{datasets}
\keyword{short}
