% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_dream_quest_of_unknown_kadath}
\alias{the_dream_quest_of_unknown_kadath}
\title{\emph{The Dream Quest of Unknown Kadath}}
\format{
A character vector with 3,475 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_dream_quest_of_unknown_kadath}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_dream_quest_of_unknown_kadath.rda}{RDA file}
}
}
\usage{
data(the_dream_quest_of_unknown_kadath)
}
\description{
A dataset containing text of H. P. Lovecraft's \emph{The Dream Quest of Unknown Kadath}.
}
\keyword{datasets}
\keyword{novella}
