% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_dream_quest_of_unknown_kadath}
\alias{the_dream_quest_of_unknown_kadath}
\title{Dataset of "THE DREAM QUEST OF UNKNOWN KADATH"}
\format{
A character vector with 3,475 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_dream_quest_of_unknown_kadath
}
\description{
A dataset containing text of H. P. Lovecraft's "THE DREAM QUEST OF UNKNOWN KADATH"
}
\examples{
data("the_dream_quest_of_unknown_kadath")
lovecraftr::the_dream_quest_of_unknown_kadath

}
\keyword{datasets}
