% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_dunwich_horror}
\alias{the_dunwich_horror}
\title{Dataset of "The Dunwich Horror"}
\format{
A character vector with 1,533 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_dunwich_horror
}
\description{
A dataset containing text of H. P. Lovecraft's "The Dunwich Horror"
}
\examples{
data("the_dunwich_horror")
lovecraftr::the_dunwich_horror

}
\keyword{datasets}
