% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_evil_clergyman}
\alias{the_evil_clergyman}
\title{"The Evil Clergyman"}
\format{
A character vector with 147 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_evil_clergyman
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Evil Clergyman"
}
\examples{
data("the_evil_clergyman")
lovecraftr::the_evil_clergyman

}
\keyword{datasets}
