% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_festival}
\alias{the_festival}
\title{Dataset of "THE FESTIVAL"}
\format{
A character vector with 303 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_festival
}
\description{
A dataset containing text of H. P. Lovecraft's "THE FESTIVAL"
}
\examples{
data("the_festival")
lovecraftr::the_festival

}
\keyword{datasets}
