% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_festival}
\alias{the_festival}
\title{Dataset of "THE FESTIVAL"}
\format{
A character vector with 303 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_festival}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_festival.rda}{RDA file}.
}
\usage{
data(the_festival)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE FESTIVAL"
}
\keyword{datasets}
