% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_haunter_of_the_dark}
\alias{the_haunter_of_the_dark}
\title{Dataset of "THE HAUNTER OF THE DARK"}
\format{
A character vector with 828 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_haunter_of_the_dark}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_haunter_of_the_dark.rda}{RDA file}.
}
\usage{
data(the_haunter_of_the_dark)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE HAUNTER OF THE DARK"
}
\keyword{datasets}
