% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_horror_at_martins_beach}
\alias{the_horror_at_martins_beach}
\title{"The Horror at Martin's Beach"}
\format{
A character vector with 225 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_horror_at_martins_beach}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_horror_at_martins_beach.rda}{RDA file}
}
}
\usage{
data(the_horror_at_martins_beach)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Horror at Martin's Beach".
}
\keyword{datasets}
\keyword{short}
