% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_horror_at_martins_beach}
\alias{the_horror_at_martins_beach}
\title{Dataset of "THE HORROR AT MARTIN'S BEACH"}
\format{
A character vector with 225 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_horror_at_martins_beach
}
\description{
A dataset containing text of H. P. Lovecraft's "THE HORROR AT MARTIN'S BEACH"
}
\examples{
data("the_horror_at_martins_beach")
lovecraftr::the_horror_at_martins_beach

}
\keyword{datasets}
