% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_horror_at_redhook}
\alias{the_horror_at_redhook}
\title{Dataset of "THE HORROR AT RED HOOK"}
\format{
A character vector with 702 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_horror_at_redhook}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_horror_at_redhook.rda}{RDA file}.
}
\usage{
data(the_horror_at_redhook)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE HORROR AT RED HOOK"
}
\keyword{datasets}
