% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_horror_at_redhook}
\alias{the_horror_at_redhook}
\title{"The Horror At Red Hook"}
\format{
A character vector with 702 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_horror_at_redhook
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Horror At Red Hook"
}
\examples{
data("the_horror_at_redhook")
lovecraftr::the_horror_at_redhook

}
\keyword{datasets}
