% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_lurking_fear}
\alias{the_lurking_fear}
\title{"The Lurking Fear"}
\format{
A character vector with 656 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_lurking_fear}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_lurking_fear.rda}{RDA file}.
}
\usage{
data(the_lurking_fear)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Lurking Fear".
}
\keyword{datasets}
\keyword{short}
