% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_lurking_fear}
\alias{the_lurking_fear}
\title{Dataset of "THE LURKING FEAR"}
\format{
A character vector with 656 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_lurking_fear
}
\description{
A dataset containing text of H. P. Lovecraft's "THE LURKING FEAR"
}
\examples{
data("the_lurking_fear")
lovecraftr::the_lurking_fear

}
\keyword{datasets}
