% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_moon_bog}
\alias{the_moon_bog}
\title{"The Moon Bog"}
\format{
A character vector with 266 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_moon_bog}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_moon_bog.rda}{RDA file}.
}
\usage{
data(the_moon_bog)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Moon Bog".
}
\keyword{datasets}
\keyword{short}
