% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_music_of_erich_zann}
\alias{the_music_of_erich_zann}
\title{Dataset of "The Music of Erich Zann"}
\format{
A character vector with 285 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_music_of_erich_zann}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_music_of_erich_zann.rda}{RDA file}.
}
\usage{
data(the_music_of_erich_zann)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Music of Erich Zann"
}
\keyword{datasets}
\keyword{short}
