% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_music_of_erich_zann}
\alias{the_music_of_erich_zann}
\title{Dataset of "THE MUSIC OF ERICH ZANN"}
\format{
A character vector with 285 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_music_of_erich_zann
}
\description{
A dataset containing text of H. P. Lovecraft's "THE MUSIC OF ERICH ZANN"
}
\examples{
data("the_music_of_erich_zann")
lovecraftr::the_music_of_erich_zann

}
\keyword{datasets}
