% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_nameless_city}
\alias{the_nameless_city}
\title{"The Nameless City"}
\format{
A character vector with 418 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_nameless_city
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Nameless City"
}
\examples{
data("the_nameless_city")
lovecraftr::the_nameless_city

}
\keyword{datasets}
