% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_nameless_city}
\alias{the_nameless_city}
\title{"The Nameless City"}
\format{
A character vector with 418 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_nameless_city}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_nameless_city.rda}{RDA file}.
}
\usage{
data(the_nameless_city)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Nameless City".
}
\keyword{datasets}
\keyword{short}
