% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_other_gods}
\alias{the_other_gods}
\title{"The Other Gods"}
\format{
A character vector with 160 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_other_gods}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_other_gods.rda}{RDA file}
}
}
\usage{
data(the_other_gods)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Other Gods".
}
\keyword{datasets}
\keyword{short}
