% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_other_gods}
\alias{the_other_gods}
\title{Dataset of "The Other Gods"}
\format{
A character vector with 160 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_other_gods
}
\description{
A dataset containing text of H. P. Lovecraft's "The Other Gods"
}
\examples{
data("the_other_gods")
lovecraftr::the_other_gods

}
\keyword{datasets}
