% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_outsider}
\alias{the_outsider}
\title{Dataset of "THE OUTSIDER"}
\format{
A character vector with 213 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_outsider
}
\description{
A dataset containing text of H. P. Lovecraft's "THE OUTSIDER"
}
\examples{
data("the_outsider")
lovecraftr::the_outsider

}
\keyword{datasets}
