% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_outsider}
\alias{the_outsider}
\title{"The Outsider"}
\format{
A character vector with 213 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_outsider
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Outsider"
}
\examples{
data("the_outsider")
lovecraftr::the_outsider

}
\keyword{datasets}
