% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_picture_in_the_house}
\alias{the_picture_in_the_house}
\title{Dataset of "The Picture In The House"}
\format{
A character vector with 263 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_picture_in_the_house
}
\description{
A dataset containing text of H. P. Lovecraft's "The Picture In The House"
}
\examples{
data("the_picture_in_the_house")
lovecraftr::the_picture_in_the_house

}
\keyword{datasets}
