% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_picture_in_the_house}
\alias{the_picture_in_the_house}
\title{Dataset of "The Picture in the House"}
\format{
A character vector with 263 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_picture_in_the_house}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_picture_in_the_house.rda}{RDA file}.
}
\usage{
data(the_picture_in_the_house)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Picture in the House"
}
\keyword{datasets}
\keyword{short}
