% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_quest_of_iranon}
\alias{the_quest_of_iranon}
\title{Dataset of "The Quest Of Iranon"}
\format{
A character vector with 215 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_quest_of_iranon
}
\description{
A dataset containing text of H. P. Lovecraft's "The Quest Of Iranon"
}
\examples{
data("the_quest_of_iranon")
lovecraftr::the_quest_of_iranon

}
\keyword{datasets}
