% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_quest_of_iranon}
\alias{the_quest_of_iranon}
\title{Dataset of "THE QUEST OF IRANON"}
\format{
A character vector with 215 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_quest_of_iranon}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_quest_of_iranon.rda}{RDA file}.
}
\usage{
data(the_quest_of_iranon)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE QUEST OF IRANON"
}
\keyword{datasets}
