% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_rats_in_the_walls}
\alias{the_rats_in_the_walls}
\title{Dataset of "THE RATS IN THE WALLS"}
\format{
A character vector with 691 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_rats_in_the_walls}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_rats_in_the_walls.rda}{RDA file}.
}
\usage{
data(the_rats_in_the_walls)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE RATS IN THE WALLS"
}
\keyword{datasets}
