% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_shadow_out_of_time}
\alias{the_shadow_out_of_time}
\title{Dataset of "The Shadow Out Of Time"}
\format{
A character vector with 2,529 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_shadow_out_of_time
}
\description{
A dataset containing text of H. P. Lovecraft's "The Shadow Out Of Time"
}
\examples{
data("the_shadow_out_of_time")
lovecraftr::the_shadow_out_of_time

}
\keyword{datasets}
