% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_shadow_over_innsmouth}
\alias{the_shadow_over_innsmouth}
\title{Dataset of \emph{The Shadow Over Innsmouth}}
\format{
A character vector with 2,318 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_shadow_over_innsmouth}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_shadow_over_innsmouth.rda}{RDA file}.
}
\usage{
data(the_shadow_over_innsmouth)
}
\description{
A dataset containing text of H. P. Lovecraft's \emph{The Shadow Over Innsmouth}
}
\keyword{datasets}
\keyword{novella}
