% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_shadow_over_innsmouth}
\alias{the_shadow_over_innsmouth}
\title{Dataset of "THE SHADOW OVER INNSMOUTH"}
\format{
A character vector with 2,318 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_shadow_over_innsmouth}{raw text file} or download
the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_shadow_over_innsmouth.rda}{RDA file} to use with R.
}
\usage{
data(the_shadow_over_innsmouth)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE SHADOW OVER INNSMOUTH"
}
\examples{
\dontrun{
library(lovecraftr)
data(the_shadow_over_innsmouth)

the_shadow_over_innsmouth <- lovecraftr::the_shadow_over_innsmouth
}
}
\keyword{datasets}
