% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_statement_of_randolph_carter}
\alias{the_statement_of_randolph_carter}
\title{Dataset of "The Statement of Randolph Carter"}
\format{
A character vector with 221 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_statement_of_randolph_carter}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_statement_of_randolph_carter.rda}{RDA file}.
}
\usage{
data(the_statement_of_randolph_carter)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Statement of Randolph Carter"
}
\keyword{datasets}
\keyword{short}
