% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_strange_high_house_in_the_mist}
\alias{the_strange_high_house_in_the_mist}
\title{"The Strange High House In The Mist"}
\format{
A character vector with 289 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_strange_high_house_in_the_mist
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Strange High House In The Mist"
}
\examples{
data("the_strange_high_house_in_the_mist")
lovecraftr::the_strange_high_house_in_the_mist

}
\keyword{datasets}
