% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_strange_high_house_in_the_mist}
\alias{the_strange_high_house_in_the_mist}
\title{Dataset of "The Strange High House in the Mist"}
\format{
A character vector with 289 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_strange_high_house_in_the_mist}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_strange_high_house_in_the_mist.rda}{RDA file}.
}
\usage{
data(the_strange_high_house_in_the_mist)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Strange High House in the Mist"
}
\keyword{datasets}
\keyword{short}
