% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_street}
\alias{the_street}
\title{Dataset of "THE STREET"}
\format{
A character vector with 183 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_street
}
\description{
A dataset containing text of H. P. Lovecraft's "THE STREET"
}
\examples{
data("the_street")
lovecraftr::the_street

}
\keyword{datasets}
