% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_street}
\alias{the_street}
\title{Dataset of "The Street"}
\format{
A character vector with 183 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_street}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_street.rda}{RDA file}.
}
\usage{
data(the_street)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Street"
}
\keyword{datasets}
\keyword{short}
