% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_street}
\alias{the_street}
\title{"The Street"}
\format{
A character vector with 183 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_street
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Street"
}
\examples{
data("the_street")
lovecraftr::the_street

}
\keyword{datasets}
