% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_temple}
\alias{the_temple}
\title{"The Temple"}
\format{
A character vector with 450 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_temple}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_temple.rda}{RDA file}
}
}
\usage{
data(the_temple)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Temple".
}
\keyword{datasets}
\keyword{short}
