% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_terrible_old_man}
\alias{the_terrible_old_man}
\title{Dataset of "THE TERRIBLE OLD MAN"}
\format{
A character vector with 87 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_terrible_old_man}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_terrible_old_man.rda}{RDA file}.
}
\usage{
data(the_terrible_old_man)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE TERRIBLE OLD MAN"
}
\keyword{datasets}
