% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_terrible_old_man}
\alias{the_terrible_old_man}
\title{"The Terrible Old Man"}
\format{
A character vector with 87 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_terrible_old_man}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_terrible_old_man.rda}{RDA file}
}
}
\usage{
data(the_terrible_old_man)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Terrible Old Man".
}
\keyword{datasets}
\keyword{short}
