% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_thing_on_the_doorstep}
\alias{the_thing_on_the_doorstep}
\title{Dataset of "THE THING ON THE DOORSTEP"}
\format{
A character vector with 931 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_thing_on_the_doorstep}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_thing_on_the_doorstep.rda}{RDA file}.
}
\usage{
data(the_thing_on_the_doorstep)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE THING ON THE DOORSTEP"
}
\keyword{datasets}
