% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_thing_on_the_doorstep}
\alias{the_thing_on_the_doorstep}
\title{"The Thing on the Doorstep"}
\format{
A character vector with 931 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_thing_on_the_doorstep}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_thing_on_the_doorstep.rda}{RDA file}
}
}
\usage{
data(the_thing_on_the_doorstep)
}
\description{
A dataset containing text of H. P. Lovecraft's "The Thing on the Doorstep".
}
\keyword{datasets}
\keyword{short}
