% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_tomb}
\alias{the_tomb}
\title{Dataset of "The Tomb"}
\format{
A character vector with 366 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_tomb
}
\description{
A dataset containing text of H. P. Lovecraft's "The Tomb"
}
\examples{
data("the_tomb")
lovecraftr::the_tomb

}
\keyword{datasets}
