% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_tomb}
\alias{the_tomb}
\title{Dataset of "THE TOMB"}
\format{
A character vector with 366 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_tomb}{raw text file} or download
the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_tomb.rda}{RDA file} to use with R.
}
\usage{
data(the_tomb)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE TOMB"
}
\examples{
\dontrun{
library(lovecraftr)
data(the_tomb)

the_tomb <- lovecraftr::the_tomb
}
}
\keyword{datasets}
