% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_transition_of_juan_romero}
\alias{the_transition_of_juan_romero}
\title{Dataset of "THE TRANSITION OF JUAN ROMERO"}
\format{
A character vector with 234 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_transition_of_juan_romero}{raw text file} or download
the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_transition_of_juan_romero.rda}{RDA file} to use with R.
}
\usage{
data(the_transition_of_juan_romero)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE TRANSITION OF JUAN ROMERO"
}
\examples{
\dontrun{
library(lovecraftr)
data(the_transition_of_juan_romero)

the_transition_of_juan_romero <- lovecraftr::the_transition_of_juan_romero
}
}
\keyword{datasets}
