% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_transition_of_juan_romero}
\alias{the_transition_of_juan_romero}
\title{"The Transition Of Juan Romero"}
\format{
A character vector with 234 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_transition_of_juan_romero
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Transition Of Juan Romero"
}
\examples{
data("the_transition_of_juan_romero")
lovecraftr::the_transition_of_juan_romero

}
\keyword{datasets}
