% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_tree}
\alias{the_tree}
\title{Dataset of "THE TREE"}
\format{
A character vector with 127 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_tree
}
\description{
A dataset containing text of H. P. Lovecraft's "THE TREE"
}
\examples{
data("the_tree")
lovecraftr::the_tree

}
\keyword{datasets}
