% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_unnamable}
\alias{the_unnamable}
\title{Dataset of "THE UNNAMABLE"}
\format{
A character vector with 239 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_unnamable}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_unnamable.rda}{RDA file}.
}
\usage{
data(the_unnamable)
}
\description{
A dataset containing text of H. P. Lovecraft's "THE UNNAMABLE"
}
\keyword{datasets}
