% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_very_old_folk}
\alias{the_very_old_folk}
\title{Dataset of "The Very Old Folk"}
\format{
A character vector with 210 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_very_old_folk
}
\description{
A dataset containing text of H. P. Lovecraft's "The Very Old Folk"
}
\examples{
data("the_very_old_folk")
lovecraftr::the_very_old_folk

}
\keyword{datasets}
