% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_whisperer_in_darkness}
\alias{the_whisperer_in_darkness}
\title{\emph{The Whisperer in Darkness}}
\format{
A character vector with 2,303 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_whisperer_in_darkness}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_whisperer_in_darkness.rda}{RDA file}.
}
\usage{
data(the_whisperer_in_darkness)
}
\description{
A dataset containing text of H. P. Lovecraft's \emph{The Whisperer in Darkness}.
}
\keyword{datasets}
\keyword{novella}
