% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_whisperer_in_darkness}
\alias{the_whisperer_in_darkness}
\title{"The Whisperer In Darkness"}
\format{
A character vector with 2,303 elements,
  with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_whisperer_in_darkness
}
\description{
A dataset containing UTF-8 plain text of
  H. P. Lovecraft's "The Whisperer In Darkness"
}
\examples{
data("the_whisperer_in_darkness")
lovecraftr::the_whisperer_in_darkness

}
\keyword{datasets}
