% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_white_ship}
\alias{the_white_ship}
\title{Dataset of "The White Ship"}
\format{
A character vector with 200 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
the_white_ship
}
\description{
A dataset containing text of H. P. Lovecraft's "The White Ship"
}
\examples{
data("the_white_ship")
lovecraftr::the_white_ship

}
\keyword{datasets}
