% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_white_ship}
\alias{the_white_ship}
\title{"The White Ship"}
\format{
A character vector with 200 elements, with roughly 80 characters per line.
}
\source{
Public domain.
For more details see the vignette:
\code{vignette("copyright-status", package = "lovecraftr")}.
\itemize{
\item \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/the_white_ship}{raw text file}
\item \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/the_white_ship.rda}{RDA file}
}
}
\usage{
data(the_white_ship)
}
\description{
A dataset containing text of H. P. Lovecraft's "The White Ship".
}
\keyword{datasets}
\keyword{short}
