% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{through_the_gates_of_the_silver_key}
\alias{through_the_gates_of_the_silver_key}
\title{"Through the Gates of the Silver Key"}
\format{
A character vector with 1,305 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/through_the_gates_of_the_silver_key}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/through_the_gates_of_the_silver_key.rda}{RDA file}.
}
\usage{
data(through_the_gates_of_the_silver_key)
}
\description{
A dataset containing text of H. P. Lovecraft's "Through the Gates of the Silver Key".
}
\keyword{datasets}
\keyword{short}
