% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{what_the_moon_brings}
\alias{what_the_moon_brings}
\title{Dataset of "What the Moon Brings"}
\format{
A character vector with 60 elements, with roughly 80 characters per line.
}
\source{
Public domain.
See the \href{https://raw.githubusercontent.com/jrdnbradford/lovecraftr/main/data-raw/corpus/what_the_moon_brings}{raw text file}
or download the \href{https://github.com/jrdnbradford/lovecraftr/raw/main/data/what_the_moon_brings.rda}{RDA file}.
}
\usage{
data(what_the_moon_brings)
}
\description{
A dataset containing text of H. P. Lovecraft's "What the Moon Brings"
}
\keyword{datasets}
\keyword{short}
