% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{what_the_moon_brings}
\alias{what_the_moon_brings}
\title{Dataset of "WHAT THE MOON BRINGS"}
\format{
A character vector with 60 elements, with roughly 80 characters per line
}
\source{
Public domain
}
\usage{
what_the_moon_brings
}
\description{
A dataset containing text of H. P. Lovecraft's "WHAT THE MOON BRINGS"
}
\examples{
data("what_the_moon_brings")
lovecraftr::what_the_moon_brings

}
\keyword{datasets}
